Create tests/test_checksum.py
