+++ tests/test_ssds.py
@@ -7,9 +7,6 @@ import logging
 import unittest
 import tempfile
 from uuid import uuid4
-from random import randint
-
-from google.cloud import storage
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
@@ -197,36 +194,6 @@ class TestSSDS(infra.SuppressWarningsMixin, unittest.TestCase):
             with open(os.path.join(root, "large.dat"), "wb") as fh:
                 fh.write(TestData.multipart())
         return root
-
-class TestSSDSChecksum(infra.SuppressWarningsMixin, unittest.TestCase):
-    def test_crc32c(self):
-        data = b"\x89\xc0\xc6\xcd\xa9$=\xfa\x91\x86\xedi\xec\x18\xcc\xad\xd1\xe1\x82\x8f^\xd2\xdd$\x1fE\x821"
-        expected_crc32c = "25c7a879"
-        with self.subTest("all at once"):
-            cs = ssds.checksum.crc32c(data)
-            self.assertEqual(expected_crc32c, cs.hexdigest())
-        with self.subTest("sliced"):
-            i = randint(1, len(data) - 2)
-            cs = ssds.checksum.crc32c(data[:i])
-            cs.update(data[i:])
-            self.assertEqual(expected_crc32c, cs.hexdigest())
-
-    def test_blob_crc32c(self):
-        data = TestData.oneshot()
-        blob = storage.Client().bucket(GS_SSDS.bucket).blob("test")
-        with io.BytesIO(data) as fh:
-            blob.upload_from_file(fh)
-        blob.reload()
-        cs = ssds.checksum.crc32c(data).google_storage_crc32c()
-        self.assertEqual(blob.crc32c, cs)
-
-    def test_blob_md5(self):
-        data = TestData.oneshot()
-        blob = ssds.aws.resource("s3").Bucket(S3_SSDS.bucket).Object("test")
-        with io.BytesIO(data) as fh:
-            blob.upload_fileobj(fh)
-        cs = ssds.checksum.md5(data).hexdigest()
-        self.assertEqual(blob.e_tag.replace('"', ''), cs)
 
 if __name__ == '__main__':
     unittest.main()
